requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
beautifulsoup4==4.12.3
pydantic==2.7.4
